refactor: render, view 통합 및 중복 코드 제거, fetch흐름 controller로 이관
